added fixed anchors on the same side
